+++ src/lib.rs
@@ -108,7 +108,6 @@ pub mod error;
 pub mod execution_result;
 pub mod executor;
 mod ffi;
-pub mod invoke;
 mod jit_runner;
 pub mod libfuncs;
 pub mod metadata;
