+++ src/values.rs
@@ -1,114 +1,598 @@
 //! # JIT params and return values de/serialization
 
+//! A Rusty interface to provide parameters to JIT calls.
+
+use std::{alloc::Layout, collections::HashMap, ops::Neg, ptr::NonNull};
+
 use bumpalo::Bump;
+use cairo_felt::Felt252;
 use cairo_lang_sierra::{
     extensions::{
         core::{CoreLibfunc, CoreType, CoreTypeConcrete},
         starknet::StarkNetTypeConcrete,
-        GenericLibfunc, GenericType,
     },
     ids::ConcreteTypeId,
     program_registry::ProgramRegistry,
 };
-use serde::{de::DeserializeSeed, Serialize};
-use std::{error::Error, fmt, ptr::NonNull};
+use num_bigint::{BigInt, Sign};
 
-pub mod array;
-pub mod bitwise;
-pub mod r#box;
-pub mod builtin_costs;
-pub mod ec_op;
-pub mod ec_point;
-pub mod ec_state;
-pub mod r#enum;
-pub mod felt252;
-pub mod felt252_dict;
-pub mod felt252_dict_entry;
-pub mod gas_builtin;
-pub mod non_zero;
-pub mod nullable;
-pub mod pedersen;
-pub mod poseidon;
-pub mod range_check;
-pub mod segment_arena;
-pub mod snapshot;
-pub mod squashed_felt252_dict;
-pub mod stark_net;
-pub mod r#struct;
-pub mod uint128;
-pub mod uint128_mul_guarantee;
-pub mod uint16;
-pub mod uint32;
-pub mod uint64;
-pub mod uint8;
-pub mod uninitialized;
+use crate::{
+    metadata::syscall_handler::SyscallHandlerMeta,
+    types::{felt252::PRIME, TypeBuilder},
+    utils::{
+        felt252_bigint, get_integer_layout, layout_repeat, next_multiple_of_usize, u32_vec_to_felt,
+    },
+};
+
+/// A JITValue is a value that can be passed to the JIT engine as an argument or received as a result.
+///
+/// They map to the cairo/sierra types.
+///
+/// The debug_name field on some variants is `Some` when receiving a [`JITValue`] as a result.
+#[derive(Educe, Debug, Clone, Eq)]
+#[educe(PartialEq)]
+pub enum JITValue {
+    Felt252(Felt252),
+    Array(Vec<Self>), // all elements need to be same type
+    Struct {
+        fields: Vec<Self>,
+        #[educe(PartialEq(ignore))]
+        debug_name: Option<String>,
+    }, // element types can differ
+    Enum {
+        tag: usize,
+        value: Box<Self>,
+        #[educe(PartialEq(ignore))]
+        debug_name: Option<String>,
+    },
+    Felt252Dict {
+        value: HashMap<Felt252, Self>,
+        #[educe(PartialEq(ignore))]
+        debug_name: Option<String>,
+    },
+    Box(Box<Self>), // can't be null
+    Nullable(Option<Box<Self>>),
+    Uint8(u8),
+    Uint16(u16),
+    Uint32(u32),
+    Uint64(u64),
+    Uint128(u128),
+    EcPoint(Felt252, Felt252),
+}
+
+#[derive(Debug, Default)]
+pub struct InvokeContext<'s> {
+    pub gas: Option<u128>,
+    // Starknet syscall handler
+    pub system: Option<&'s SyscallHandlerMeta>,
+    // call args
+    pub args: Vec<JITValue>,
+}
+
+#[derive(Debug, Default)]
+pub struct InvokeResult {
+    pub gas: Option<u128>,
+    pub outputs: Vec<JITValue>,
+}
+
+// Conversions
+
+impl From<Felt252> for JITValue {
+    fn from(value: Felt252) -> Self {
+        JITValue::Felt252(value)
+    }
+}
+
+impl From<u8> for JITValue {
+    fn from(value: u8) -> Self {
+        JITValue::Uint8(value)
+    }
+}
+
+impl From<u16> for JITValue {
+    fn from(value: u16) -> Self {
+        JITValue::Uint16(value)
+    }
+}
+
+impl From<u32> for JITValue {
+    fn from(value: u32) -> Self {
+        JITValue::Uint32(value)
+    }
+}
+
+impl From<u64> for JITValue {
+    fn from(value: u64) -> Self {
+        JITValue::Uint64(value)
+    }
+}
+
+impl From<u128> for JITValue {
+    fn from(value: u128) -> Self {
+        JITValue::Uint128(value)
+    }
+}
+
+impl JITValue {
+    pub(crate) fn resolve_type<'a>(
+        ty: &'a CoreTypeConcrete,
+        registry: &'a ProgramRegistry<CoreType, CoreLibfunc>,
+    ) -> &'a CoreTypeConcrete {
+        match ty {
+            CoreTypeConcrete::Snapshot(info) => registry.get_type(&info.ty).unwrap(),
+            x => x,
+        }
+    }
+
+    /// Allocates the value in the given arena so it can be passed to the JIT engine.
+    pub(crate) fn to_jit(
+        &self,
+        arena: &Bump,
+        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+        type_id: &ConcreteTypeId,
+    ) -> NonNull<()> {
+        let ty = registry.get_type(type_id).unwrap();
+
+        unsafe {
+            match self {
+                JITValue::Felt252(value) => {
+                    let ptr = arena.alloc_layout(get_integer_layout(252)).cast();
+
+                    let data = felt252_bigint(value.to_bigint());
+                    ptr.cast::<[u32; 8]>().as_mut().copy_from_slice(&data);
+                    ptr
+                }
+                JITValue::Array(data) => {
+                    if let CoreTypeConcrete::Array(info) = Self::resolve_type(ty, registry) {
+                        // todo: if its snapshot  cargo r --example starknet
+                        let elem_ty = registry.get_type(&info.ty).unwrap();
+                        let elem_layout = elem_ty.layout(registry).unwrap().pad_to_align();
+
+                        let ptr: *mut NonNull<()> =
+                            libc::malloc(elem_layout.size() * data.len()).cast();
+                        let mut len: u32 = 0;
+                        let cap: u32 = data.len().try_into().unwrap();
+
+                        for elem in data {
+                            let elem = elem.to_jit(arena, registry, &info.ty);
+
+                            std::ptr::copy_nonoverlapping(
+                                elem.cast::<u8>().as_ptr(),
+                                NonNull::new(
+                                    ((NonNull::new_unchecked(ptr).as_ptr() as usize)
+                                        + len as usize * elem_layout.size())
+                                        as *mut u8,
+                                )
+                                .unwrap()
+                                .cast()
+                                .as_ptr(),
+                                elem_layout.size(),
+                            );
+
+                            len += 1;
+                        }
+
+                        let target = arena.alloc_layout(
+                            Layout::new::<*mut NonNull<()>>()
+                                .extend(Layout::new::<u32>())
+                                .unwrap()
+                                .0
+                                .extend(Layout::new::<u32>())
+                                .unwrap()
+                                .0,
+                        );
+
+                        *target.cast::<*mut NonNull<()>>().as_mut() = ptr;
+
+                        let (layout, offset) = Layout::new::<*mut NonNull<()>>()
+                            .extend(Layout::new::<u32>())
+                            .unwrap();
+
+                        *NonNull::new(((target.as_ptr() as usize) + offset) as *mut u32)
+                            .unwrap()
+                            .cast()
+                            .as_mut() = len;
+
+                        let (_, offset) = layout.extend(Layout::new::<u32>()).unwrap();
+
+                        *NonNull::new(((target.as_ptr() as usize) + offset) as *mut u32)
+                            .unwrap()
+                            .cast()
+                            .as_mut() = cap;
+                        target.cast()
+                    } else {
+                        panic!("wrong type: {:?}", type_id.debug_name.as_ref())
+                    }
+                }
+                JITValue::Struct {
+                    fields: members, ..
+                } => {
+                    if let CoreTypeConcrete::Struct(info) = Self::resolve_type(ty, registry) {
+                        let mut layout: Option<Layout> = None;
+                        let mut data = Vec::with_capacity(info.members.len());
+
+                        for (member_type_id, member) in info.members.iter().zip(members) {
+                            let member_ty = registry.get_type(member_type_id).unwrap();
+                            let member_layout = member_ty.layout(registry).unwrap();
+
+                            let (new_layout, offset) = match layout {
+                                Some(layout) => layout.extend(member_layout).unwrap(),
+                                None => (member_layout, 0),
+                            };
+                            layout = Some(new_layout);
+
+                            data.push((
+                                member_layout,
+                                offset,
+                                member.to_jit(arena, registry, member_type_id),
+                            ));
+                        }
+
+                        let ptr = arena
+                            .alloc_layout(layout.unwrap_or(Layout::new::<()>()))
+                            .cast();
+
+                        for (layout, offset, member_ptr) in data {
+                            std::ptr::copy_nonoverlapping(
+                                member_ptr.cast::<u8>().as_ptr(),
+                                NonNull::new(((ptr.as_ptr() as usize) + offset) as *mut u8)
+                                    .unwrap()
+                                    .cast()
+                                    .as_ptr(),
+                                layout.size(),
+                            );
+                        }
+
+                        ptr
+                    } else {
+                        panic!("wrong type")
+                    }
+                }
+                JITValue::Enum { tag, value, .. } => {
+                    if let CoreTypeConcrete::Enum(info) = Self::resolve_type(ty, registry) {
+                        let tag_value = *tag;
+                        assert!(tag_value <= info.variants.len());
+
+                        let payload_type_id = &info.variants[tag_value];
+                        let payload = value.to_jit(arena, registry, payload_type_id);
+
+                        let (layout, tag_layout, variant_layouts) =
+                            crate::types::r#enum::get_layout_for_variants(registry, &info.variants)
+                                .unwrap();
+                        let ptr = arena.alloc_layout(layout).cast();
+
+                        match tag_layout.size() {
+                            1 => *ptr.cast::<u8>().as_mut() = tag_value as u8,
+                            2 => *ptr.cast::<u16>().as_mut() = tag_value as u16,
+                            4 => *ptr.cast::<u32>().as_mut() = tag_value as u32,
+                            8 => *ptr.cast::<u64>().as_mut() = tag_value as u64,
+                            _ => unreachable!(),
+                        }
+
+                        std::ptr::copy_nonoverlapping(
+                            payload.cast::<u8>().as_ptr(),
+                            NonNull::new(
+                                ((ptr.as_ptr() as usize)
+                                    + tag_layout.extend(variant_layouts[tag_value]).unwrap().1)
+                                    as *mut u8,
+                            )
+                            .unwrap()
+                            .cast()
+                            .as_ptr(),
+                            variant_layouts[tag_value].size(),
+                        );
+
+                        ptr
+                    } else {
+                        panic!("wrong type")
+                    }
+                }
+                JITValue::Felt252Dict { value: map, .. } => {
+                    if let CoreTypeConcrete::Felt252Dict(info) = Self::resolve_type(ty, registry) {
+                        let elem_ty = registry.get_type(&info.ty).unwrap();
+                        let elem_layout = elem_ty.layout(registry).unwrap().pad_to_align();
+
+                        let mut value_map = HashMap::<[u8; 32], NonNull<std::ffi::c_void>>::new();
+
+                        // next key must be called before next_value
+
+                        for (key, value) in map.iter() {
+                            let key = key.to_le_bytes();
+                            let value = value.to_jit(arena, registry, &info.ty);
+
+                            let value_malloc_ptr =
+                                NonNull::new(libc::malloc(elem_layout.size())).unwrap();
+
+                            std::ptr::copy_nonoverlapping(
+                                value.cast::<u8>().as_ptr(),
+                                value_malloc_ptr.cast().as_ptr(),
+                                elem_layout.size(),
+                            );
+
+                            value_map.insert(key, value_malloc_ptr);
+                        }
+
+                        let target: NonNull<NonNull<HashMap<[u8; 32], NonNull<std::ffi::c_void>>>> =
+                            arena
+                                .alloc_layout(Layout::new::<
+                                    NonNull<HashMap<[u8; 32], NonNull<std::ffi::c_void>>>,
+                                >())
+                                .cast();
+
+                        let map_ptr: NonNull<HashMap<[u8; 32], NonNull<std::ffi::c_void>>> = arena
+                            .alloc_layout(
+                                Layout::new::<HashMap<[u8; 32], NonNull<std::ffi::c_void>>>()
+                                    .pad_to_align(),
+                            )
+                            .cast();
+
+                        std::ptr::write(map_ptr.as_ptr(), value_map);
+                        std::ptr::write(target.as_ptr(), map_ptr);
+
+                        target.cast()
+                    } else {
+                        panic!("wrong type")
+                    }
+                }
+                JITValue::Box(_) => todo!(),
+                JITValue::Nullable(_) => todo!(),
+                JITValue::Uint8(value) => {
+                    let ptr = arena.alloc_layout(Layout::new::<u8>()).cast();
+                    *ptr.cast::<u8>().as_mut() = *value;
+
+                    ptr
+                }
+                JITValue::Uint16(value) => {
+                    let ptr = arena.alloc_layout(Layout::new::<u16>()).cast();
+                    *ptr.cast::<u16>().as_mut() = *value;
+
+                    ptr
+                }
+                JITValue::Uint32(value) => {
+                    let ptr = arena.alloc_layout(Layout::new::<u32>()).cast();
+                    *ptr.cast::<u32>().as_mut() = *value;
+
+                    ptr
+                }
+                JITValue::Uint64(value) => {
+                    let ptr = arena.alloc_layout(Layout::new::<u64>()).cast();
+                    *ptr.cast::<u64>().as_mut() = *value;
+
+                    ptr
+                }
+                JITValue::Uint128(value) => {
+                    let ptr = arena.alloc_layout(Layout::new::<u128>()).cast();
+                    *ptr.cast::<u128>().as_mut() = *value;
+
+                    ptr
+                }
+                JITValue::EcPoint(a, b) => {
+                    let ptr = arena
+                        .alloc_layout(layout_repeat(&get_integer_layout(252), 2).unwrap().0)
+                        .cast();
+
+                    let a = felt252_bigint(a.to_bigint());
+                    let b = felt252_bigint(b.to_bigint());
+                    let data = [a, b];
+
+                    ptr.cast::<[[u32; 8]; 2]>().as_mut().copy_from_slice(&data);
+
+                    ptr
+                }
+            }
+        }
+    }
+
+    /// From the given pointer acquired from the JIT outputs, convert it to a [`Self`]
+    pub(crate) fn from_jit(
+        ptr: NonNull<()>,
+        type_id: &ConcreteTypeId,
+        registry: &ProgramRegistry<CoreType, CoreLibfunc>,
+    ) -> JITValue {
+        let ty = registry.get_type(type_id).unwrap();
+
+        unsafe {
+            match ty {
+                CoreTypeConcrete::Array(info) => {
+                    let elem_ty = registry.get_type(&info.ty).unwrap();
+
+                    let elem_layout = elem_ty.layout(registry).unwrap();
+                    let elem_stride = elem_layout.pad_to_align().size();
+
+                    let ptr_layout = Layout::new::<*mut ()>();
+                    let len_layout = crate::utils::get_integer_layout(32);
+
+                    let len_value = *NonNull::new(
+                        ((ptr.as_ptr() as usize) + ptr_layout.extend(len_layout).unwrap().1)
+                            as *mut (),
+                    )
+                    .unwrap()
+                    .cast::<u32>()
+                    .as_ref();
+
+                    let data_ptr = *ptr.cast::<NonNull<()>>().as_ref();
+                    let mut array_value = Vec::new();
+
+                    for i in 0..(len_value as usize) {
+                        let cur_elem_ptr = NonNull::new(
+                            ((data_ptr.as_ptr() as usize) + elem_stride * i) as *mut (),
+                        )
+                        .unwrap();
+
+                        array_value.push(Self::from_jit(cur_elem_ptr, &info.ty, registry));
+                    }
+
+                    Self::Array(array_value)
+                }
+                CoreTypeConcrete::Box(_) => todo!(),
+                CoreTypeConcrete::EcOp(_) => todo!(),
+                CoreTypeConcrete::EcPoint(_) => {
+                    let data = ptr.cast::<[[u32; 8]; 2]>().as_ref();
+
+                    JITValue::EcPoint(u32_vec_to_felt(&data[0]), u32_vec_to_felt(&data[1]))
+                }
+                CoreTypeConcrete::EcState(_) => todo!(),
+                CoreTypeConcrete::Felt252(_) => {
+                    let data = ptr.cast::<[u32; 8]>().as_ref();
+                    let data = u32_vec_to_felt(data);
+                    JITValue::Felt252(data)
+                }
+                CoreTypeConcrete::Uint8(_) => JITValue::Uint8(*ptr.cast::<u8>().as_ref()),
+                CoreTypeConcrete::Uint16(_) => JITValue::Uint16(*ptr.cast::<u16>().as_ref()),
+                CoreTypeConcrete::Uint32(_) => JITValue::Uint32(*ptr.cast::<u32>().as_ref()),
+                CoreTypeConcrete::Uint64(_) => JITValue::Uint64(*ptr.cast::<u64>().as_ref()),
+                CoreTypeConcrete::Uint128(_) => JITValue::Uint128(*ptr.cast::<u128>().as_ref()),
+                CoreTypeConcrete::Uint128MulGuarantee(_) => todo!(),
+                CoreTypeConcrete::Sint8(_) => todo!(),
+                CoreTypeConcrete::Sint16(_) => todo!(),
+                CoreTypeConcrete::Sint32(_) => todo!(),
+                CoreTypeConcrete::Sint64(_) => todo!(),
+                CoreTypeConcrete::Sint128(_) => todo!(),
+                CoreTypeConcrete::NonZero(_) => todo!(),
+                CoreTypeConcrete::Nullable(_) => todo!(),
+                CoreTypeConcrete::Uninitialized(_) => todo!(),
+                CoreTypeConcrete::Enum(info) => {
+                    let tag_layout = crate::utils::get_integer_layout(match info.variants.len() {
+                        0 | 1 => 0,
+                        num_variants => {
+                            (next_multiple_of_usize(num_variants.next_power_of_two(), 8) >> 3)
+                                .try_into()
+                                .unwrap()
+                        }
+                    });
+                    let tag_value = match info.variants.len() {
+                        0 => {
+                            // An enum without variants is basically the `!` (never) type in Rust.
+                            panic!("An enum without variants is not a valid type.")
+                        }
+                        1 => 0,
+                        _ => match tag_layout.size() {
+                            1 => *ptr.cast::<u8>().as_ref() as usize,
+                            2 => *ptr.cast::<u16>().as_ref() as usize,
+                            4 => *ptr.cast::<u32>().as_ref() as usize,
+                            8 => *ptr.cast::<u64>().as_ref() as usize,
+                            _ => unreachable!(),
+                        },
+                    };
+
+                    let payload_ty = registry.get_type(&info.variants[tag_value]).unwrap();
+                    let payload_layout = payload_ty.layout(registry).unwrap();
+
+                    let payload_ptr = NonNull::new(
+                        ((ptr.as_ptr() as usize) + tag_layout.extend(payload_layout).unwrap().1)
+                            as *mut _,
+                    )
+                    .unwrap();
+                    let payload = Self::from_jit(payload_ptr, &info.variants[tag_value], registry);
+
+                    Self::Enum {
+                        tag: tag_value,
+                        value: Box::new(payload),
+                        debug_name: type_id.debug_name.as_ref().map(|x| x.to_string()),
+                    }
+                }
+                CoreTypeConcrete::Struct(info) => {
+                    let mut layout: Option<Layout> = None;
+                    let mut members = Vec::with_capacity(info.members.len());
+
+                    for member_ty in &info.members {
+                        let member = registry.get_type(member_ty).unwrap();
+                        let member_layout = member.layout(registry).unwrap();
+
+                        let (new_layout, offset) = match layout {
+                            Some(layout) => layout.extend(member_layout).unwrap(),
+                            None => (member_layout, 0),
+                        };
+                        layout = Some(new_layout);
+
+                        members.push(JITValue::from_jit(
+                            NonNull::new(((ptr.as_ptr() as usize) + offset) as *mut ()).unwrap(),
+                            member_ty,
+                            registry,
+                        ));
+                    }
+
+                    JITValue::Struct {
+                        fields: members,
+                        debug_name: type_id.debug_name.as_ref().map(|x| x.to_string()),
+                    }
+                }
+                CoreTypeConcrete::Felt252Dict(info)
+                | CoreTypeConcrete::SquashedFelt252Dict(info) => {
+                    let ptr = ptr.cast::<NonNull<HashMap<[u8; 32], NonNull<std::ffi::c_void>>>>();
+                    let ptr = *ptr.as_ptr();
+                    let map = Box::from_raw(ptr.as_ptr());
+
+                    let mut output_map = HashMap::with_capacity(map.len());
+
+                    for (key, val_ptr) in map.iter() {
+                        let key = Felt252::from_bytes_le(key.as_slice());
+                        output_map.insert(key, Self::from_jit(val_ptr.cast(), &info.ty, registry));
+                    }
+
+                    Box::leak(map); // we must leak to avoid a double free
+
+                    JITValue::Felt252Dict {
+                        value: output_map,
+                        debug_name: type_id.debug_name.as_ref().map(|x| x.to_string()),
+                    }
+                }
+                CoreTypeConcrete::Felt252DictEntry(_) => {
+                    unimplemented!("shouldn't be possible to return")
+                }
+                CoreTypeConcrete::Pedersen(_)
+                | CoreTypeConcrete::Poseidon(_)
+                | CoreTypeConcrete::Bitwise(_)
+                | CoreTypeConcrete::BuiltinCosts(_)
+                | CoreTypeConcrete::RangeCheck(_)
+                | CoreTypeConcrete::GasBuiltin(_)
+                | CoreTypeConcrete::SegmentArena(_) => {
+                    unimplemented!("handled before: {:?}", type_id)
+                }
+                // Does it make sense for programs to return this? Should it be implemented
+                CoreTypeConcrete::StarkNet(selector) => match selector {
+                    StarkNetTypeConcrete::ClassHash(_)
+                    | StarkNetTypeConcrete::ContractAddress(_)
+                    | StarkNetTypeConcrete::StorageBaseAddress(_)
+                    | StarkNetTypeConcrete::StorageAddress(_) => {
+                        // felt values
+                        let data = ptr.cast::<[u32; 8]>().as_ref();
+                        let data = u32_vec_to_felt(data);
+                        JITValue::Felt252(data)
+                    }
+                    StarkNetTypeConcrete::System(_) => unimplemented!("should be handled before"),
+                    StarkNetTypeConcrete::Secp256Point(_) => todo!(),
+                },
+                CoreTypeConcrete::Span(_) => todo!(),
+                CoreTypeConcrete::Snapshot(info) => Self::from_jit(ptr, &info.ty, registry),
+                CoreTypeConcrete::Bytes31(_) => todo!(),
+            }
+        }
+    }
+
+    pub fn felt_str(value: &str) -> Self {
+        let value = value.parse::<BigInt>().unwrap();
+        let value = match value.sign() {
+            Sign::Minus => &*PRIME - value.neg().to_biguint().unwrap(),
+            _ => value.to_biguint().unwrap(),
+        };
+
+        Self::Felt252(Felt252::from(value))
+    }
+}
 
 /// The [`ValueBuilder`] trait is implemented any de/serializable value, which is the `TType`
 /// generic.
-pub trait ValueBuilder<TType, TLibfunc>
-where
-    TType: GenericType<Concrete = Self>,
-    TLibfunc: GenericLibfunc,
-{
-    /// Value deserializer from [serde] into the JIT ABI.
-    type Deserializer<'a>: ValueDeserializer<'a, TType, TLibfunc>;
-    /// Value serializer from the JIT ABI into [serde].
-    type Serializer<'a>: ValueSerializer<'a, TType, TLibfunc>;
-
-    /// Error type returned from the de/serializers.
-    type Error: Error;
-
+pub trait ValueBuilder {
     /// Return whether the type is considered complex or simple.
     ///
     /// Complex types are always passed by pointer (both as params and return values) and require a
     /// stack allocation. Examples of complex values include structs and enums, but not felts since
     /// LLVM considers them integers.
     fn is_complex(&self) -> bool;
-
-    /// Write a representation of the argument in a human-friendly format for debugging.
-    unsafe fn debug_fmt(
-        &self,
-        f: &mut fmt::Formatter,
-        id: &ConcreteTypeId,
-        registry: &ProgramRegistry<TType, TLibfunc>,
-        ptr: NonNull<()>,
-    ) -> fmt::Result;
 }
 
-/// Deserialize a value from [serde] into the JIT ABI.
-pub trait ValueDeserializer<'a, TType, TLibfunc>
-where
-    Self: for<'de> DeserializeSeed<'de, Value = NonNull<()>>,
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-{
-    fn new(
-        arena: &'a Bump,
-        registry: &'a ProgramRegistry<TType, TLibfunc>,
-        info: &'a <TType as GenericType>::Concrete,
-    ) -> Self;
-}
-
-/// Serialize a value from the JIT ABI into [serde].
-pub trait ValueSerializer<'a, TType, TLibfunc>
-where
-    Self: Serialize,
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-{
-    fn new(
-        ptr: NonNull<()>,
-        registry: &'a ProgramRegistry<TType, TLibfunc>,
-        info: &'a <TType as GenericType>::Concrete,
-    ) -> Self;
-}
-
-impl ValueBuilder<CoreType, CoreLibfunc> for CoreTypeConcrete {
-    type Deserializer<'a> = CoreTypeDeserializer<'a, CoreType, CoreLibfunc>;
-    type Serializer<'a> = CoreTypeSerializer<'a, CoreType, CoreLibfunc>;
-
-    type Error = std::convert::Infallible;
-
+impl ValueBuilder for CoreTypeConcrete {
     fn is_complex(&self) -> bool {
         match self {
             CoreTypeConcrete::Array(_) => true,
@@ -148,350 +632,6 @@ impl ValueBuilder<CoreType, CoreLibfunc> for CoreTypeConcrete {
             },
             CoreTypeConcrete::SegmentArena(_) => false,
             CoreTypeConcrete::Snapshot(_) => false,
-            CoreTypeConcrete::Sint8(_) => todo!(),
-            CoreTypeConcrete::Sint16(_) => todo!(),
-            CoreTypeConcrete::Sint32(_) => todo!(),
-            CoreTypeConcrete::Sint64(_) => todo!(),
-            CoreTypeConcrete::Sint128(_) => todo!(),
-            CoreTypeConcrete::Bytes31(_) => todo!(),
-        }
-    }
-
-    unsafe fn debug_fmt(
-        &self,
-        _f: &mut fmt::Formatter,
-        _id: &ConcreteTypeId,
-        _registry: &ProgramRegistry<CoreType, CoreLibfunc>,
-        _ptr: NonNull<()>,
-    ) -> fmt::Result {
-        todo!()
-        // match self {
-        //     CoreTypeConcrete::Array(_) => todo!(),
-        //     CoreTypeConcrete::Bitwise(_) => todo!(),
-        //     CoreTypeConcrete::Box(_) => todo!(),
-        //     CoreTypeConcrete::EcOp(_) => todo!(),
-        //     CoreTypeConcrete::EcPoint(_) => todo!(),
-        //     CoreTypeConcrete::EcState(_) => todo!(),
-        //     CoreTypeConcrete::Felt252(info) => self::felt252::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::GasBuiltin(_) => todo!(),
-        //     CoreTypeConcrete::BuiltinCosts(_) => todo!(),
-        //     CoreTypeConcrete::Uint8(info) => self::uint8::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::Uint16(info) => self::uint16::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::Uint32(info) => self::uint32::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::Uint64(info) => self::uint64::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::Uint128(_) => todo!(),
-        //     CoreTypeConcrete::Uint128MulGuarantee(_) => todo!(),
-        //     CoreTypeConcrete::NonZero(_) => todo!(),
-        //     CoreTypeConcrete::Nullable(_) => todo!(),
-        //     CoreTypeConcrete::RangeCheck(_) => todo!(),
-        //     CoreTypeConcrete::Uninitialized(_) => todo!(),
-        //     CoreTypeConcrete::Enum(_) => todo!(),
-        //     CoreTypeConcrete::Struct(info) => self::r#struct::debug_fmt(f, id, registry, ptr, info),
-        //     CoreTypeConcrete::Felt252Dict(_) => todo!(),
-        //     CoreTypeConcrete::Felt252DictEntry(_) => todo!(),
-        //     CoreTypeConcrete::SquashedFelt252Dict(_) => todo!(),
-        //     CoreTypeConcrete::Pedersen(_) => todo!(),
-        //     CoreTypeConcrete::Poseidon(_) => todo!(),
-        //     CoreTypeConcrete::Span(_) => todo!(),
-        //     CoreTypeConcrete::StarkNet(_) => todo!(),
-        //     CoreTypeConcrete::SegmentArena(_) => todo!(),
-        //     CoreTypeConcrete::Snapshot(_) => todo!(),
-        // }
-    }
-}
-
-/// Deserializer for Cairo's [`CoreType`].
-pub struct CoreTypeDeserializer<'a, TType, TLibfunc>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-{
-    arena: &'a Bump,
-    registry: &'a ProgramRegistry<TType, TLibfunc>,
-    info: &'a <TType as GenericType>::Concrete,
-}
-
-impl<'a> ValueDeserializer<'a, CoreType, CoreLibfunc>
-    for CoreTypeDeserializer<'a, CoreType, CoreLibfunc>
-{
-    fn new(
-        arena: &'a Bump,
-        registry: &'a ProgramRegistry<CoreType, CoreLibfunc>,
-        info: &'a CoreTypeConcrete,
-    ) -> Self {
-        Self {
-            arena,
-            registry,
-            info,
-        }
-    }
-}
-
-impl<'a, 'de> DeserializeSeed<'de> for CoreTypeDeserializer<'a, CoreType, CoreLibfunc> {
-    type Value = NonNull<()>;
-
-    fn deserialize<D>(self, deserializer: D) -> Result<Self::Value, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        unsafe {
-            match self.info {
-                CoreTypeConcrete::Array(info) => {
-                    self::array::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Bitwise(info) => {
-                    self::bitwise::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Box(_) => todo!(),
-                CoreTypeConcrete::EcOp(info) => {
-                    self::ec_op::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::EcPoint(info) => {
-                    self::ec_point::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::EcState(info) => {
-                    self::ec_state::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Felt252(info) => {
-                    self::felt252::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::GasBuiltin(info) => {
-                    self::gas_builtin::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::BuiltinCosts(_) => todo!(),
-                CoreTypeConcrete::Uint8(info) => {
-                    self::uint8::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uint16(info) => {
-                    self::uint16::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uint32(info) => {
-                    self::uint32::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uint64(info) => {
-                    self::uint64::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uint128(info) => {
-                    self::uint128::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uint128MulGuarantee(_) => todo!(),
-                CoreTypeConcrete::NonZero(info) => {
-                    self::non_zero::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Nullable(info) => {
-                    self::nullable::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::RangeCheck(info) => {
-                    self::range_check::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Uninitialized(_) => todo!(),
-                CoreTypeConcrete::Enum(info) => {
-                    self::r#enum::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Struct(info) => {
-                    self::r#struct::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Felt252Dict(info) => {
-                    self::felt252_dict::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Felt252DictEntry(_) => todo!(),
-                CoreTypeConcrete::SquashedFelt252Dict(info) => {
-                    self::squashed_felt252_dict::deserialize(
-                        deserializer,
-                        self.arena,
-                        self.registry,
-                        info,
-                    )
-                }
-                CoreTypeConcrete::Pedersen(info) => {
-                    self::pedersen::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Poseidon(info) => {
-                    self::poseidon::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Span(_) => todo!(),
-                CoreTypeConcrete::StarkNet(selector) => match selector {
-                    StarkNetTypeConcrete::ClassHash(info) => self::stark_net::deserialize_address(
-                        deserializer,
-                        self.arena,
-                        self.registry,
-                        info,
-                    ),
-                    StarkNetTypeConcrete::ContractAddress(info) => {
-                        self::stark_net::deserialize_address(
-                            deserializer,
-                            self.arena,
-                            self.registry,
-                            info,
-                        )
-                    }
-                    StarkNetTypeConcrete::StorageBaseAddress(info) => {
-                        self::stark_net::deserialize_address(
-                            deserializer,
-                            self.arena,
-                            self.registry,
-                            info,
-                        )
-                    }
-                    StarkNetTypeConcrete::StorageAddress(info) => {
-                        self::stark_net::deserialize_address(
-                            deserializer,
-                            self.arena,
-                            self.registry,
-                            info,
-                        )
-                    }
-                    StarkNetTypeConcrete::System(info) => self::stark_net::deserialize_system(
-                        deserializer,
-                        self.arena,
-                        self.registry,
-                        info,
-                    ),
-                    StarkNetTypeConcrete::Secp256Point(_) => todo!(),
-                },
-                CoreTypeConcrete::SegmentArena(info) => {
-                    self::segment_arena::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Snapshot(info) => {
-                    self::snapshot::deserialize(deserializer, self.arena, self.registry, info)
-                }
-                CoreTypeConcrete::Sint8(_) => todo!(),
-                CoreTypeConcrete::Sint16(_) => todo!(),
-                CoreTypeConcrete::Sint32(_) => todo!(),
-                CoreTypeConcrete::Sint64(_) => todo!(),
-                CoreTypeConcrete::Sint128(_) => todo!(),
-                CoreTypeConcrete::Bytes31(_) => todo!(),
-            }
-        }
-    }
-}
-
-/// Serializer for Cairo's [`CoreType`].
-pub struct CoreTypeSerializer<'a, TType, TLibfunc>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-{
-    ptr: NonNull<()>,
-    registry: &'a ProgramRegistry<TType, TLibfunc>,
-    info: &'a <TType as GenericType>::Concrete,
-}
-
-impl<'a> ValueSerializer<'a, CoreType, CoreLibfunc>
-    for CoreTypeSerializer<'a, CoreType, CoreLibfunc>
-{
-    fn new(
-        ptr: NonNull<()>,
-        registry: &'a ProgramRegistry<CoreType, CoreLibfunc>,
-        info: &'a <CoreType as GenericType>::Concrete,
-    ) -> Self {
-        Self {
-            ptr,
-            registry,
-            info,
-        }
-    }
-}
-
-impl<'a> Serialize for CoreTypeSerializer<'a, CoreType, CoreLibfunc> {
-    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: serde::Serializer,
-    {
-        match self.info {
-            CoreTypeConcrete::Array(info) => unsafe {
-                self::array::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Bitwise(info) => unsafe {
-                self::bitwise::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Box(_) => todo!(),
-            CoreTypeConcrete::EcOp(info) => unsafe {
-                self::ec_op::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::EcPoint(info) => unsafe {
-                self::ec_point::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::EcState(info) => unsafe {
-                self::ec_state::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Felt252(info) => unsafe {
-                self::felt252::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::GasBuiltin(info) => unsafe {
-                self::gas_builtin::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::BuiltinCosts(_) => todo!(),
-            CoreTypeConcrete::Uint8(info) => unsafe {
-                self::uint8::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uint16(info) => unsafe {
-                self::uint16::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uint32(info) => unsafe {
-                self::uint32::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uint64(info) => unsafe {
-                self::uint64::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uint128(info) => unsafe {
-                self::uint128::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uint128MulGuarantee(_) => todo!(),
-            CoreTypeConcrete::NonZero(info) => unsafe {
-                self::non_zero::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Nullable(info) => unsafe {
-                self::nullable::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::RangeCheck(info) => unsafe {
-                self::range_check::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Uninitialized(_) => todo!(),
-            CoreTypeConcrete::Enum(info) => unsafe {
-                self::r#enum::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Struct(info) => unsafe {
-                self::r#struct::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Felt252Dict(info) => unsafe {
-                self::felt252_dict::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Felt252DictEntry(_) => todo!(),
-            CoreTypeConcrete::SquashedFelt252Dict(info) => unsafe {
-                self::squashed_felt252_dict::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Pedersen(info) => unsafe {
-                self::pedersen::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Poseidon(info) => unsafe {
-                self::poseidon::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Span(_) => todo!(),
-            CoreTypeConcrete::StarkNet(selector) => match selector {
-                StarkNetTypeConcrete::ClassHash(info) => unsafe {
-                    self::stark_net::serialize_address(serializer, self.registry, self.ptr, info)
-                },
-                StarkNetTypeConcrete::ContractAddress(info) => unsafe {
-                    self::stark_net::serialize_address(serializer, self.registry, self.ptr, info)
-                },
-                StarkNetTypeConcrete::StorageBaseAddress(info) => unsafe {
-                    self::stark_net::serialize_address(serializer, self.registry, self.ptr, info)
-                },
-                StarkNetTypeConcrete::StorageAddress(info) => unsafe {
-                    self::stark_net::serialize_address(serializer, self.registry, self.ptr, info)
-                },
-                StarkNetTypeConcrete::System(info) => unsafe {
-                    self::stark_net::serialize_system(serializer, self.registry, self.ptr, info)
-                },
-                StarkNetTypeConcrete::Secp256Point(_) => todo!(),
-            },
-            CoreTypeConcrete::SegmentArena(info) => unsafe {
-                self::segment_arena::serialize(serializer, self.registry, self.ptr, info)
-            },
-            CoreTypeConcrete::Snapshot(info) => unsafe {
-                self::snapshot::serialize(serializer, self.registry, self.ptr, info)
-            },
             CoreTypeConcrete::Sint8(_) => todo!(),
             CoreTypeConcrete::Sint16(_) => todo!(),
             CoreTypeConcrete::Sint32(_) => todo!(),
