--- src/values/pedersen.rs
@@ -1,40 +1,0 @@
-use super::ValueBuilder;
-use crate::types::TypeBuilder;
-use bumpalo::Bump;
-use cairo_lang_sierra::{
-    extensions::{types::InfoOnlyConcreteType, GenericLibfunc, GenericType},
-    program_registry::ProgramRegistry,
-};
-use serde::{Deserialize, Deserializer, Serialize, Serializer};
-use std::{alloc::Layout, ptr::NonNull};
-
-pub unsafe fn deserialize<'de, TType, TLibfunc, D>(
-    deserializer: D,
-    arena: &Bump,
-    _registry: &ProgramRegistry<TType, TLibfunc>,
-    _info: &InfoOnlyConcreteType,
-) -> Result<NonNull<()>, D::Error>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-    <TType as GenericType>::Concrete: TypeBuilder<TType, TLibfunc> + ValueBuilder<TType, TLibfunc>,
-    D: Deserializer<'de>,
-{
-    <() as Deserialize>::deserialize(deserializer)
-        .map(|_| arena.alloc_layout(Layout::new::<()>()).cast())
-}
-
-pub unsafe fn serialize<TType, TLibfunc, S>(
-    serializer: S,
-    _registry: &ProgramRegistry<TType, TLibfunc>,
-    _ptr: NonNull<()>,
-    _info: &InfoOnlyConcreteType,
-) -> Result<S::Ok, S::Error>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-    <TType as GenericType>::Concrete: TypeBuilder<TType, TLibfunc> + ValueBuilder<TType, TLibfunc>,
-    S: Serializer,
-{
-    <() as Serialize>::serialize(&(), serializer)
-}
