--- src/values/snapshot.rs
@@ -1,47 +1,0 @@
-use super::{ValueBuilder, ValueDeserializer, ValueSerializer};
-use crate::types::TypeBuilder;
-use bumpalo::Bump;
-use cairo_lang_sierra::{
-    extensions::{types::InfoAndTypeConcreteType, GenericLibfunc, GenericType},
-    program_registry::ProgramRegistry,
-};
-use serde::{de::DeserializeSeed, Deserializer, Serialize, Serializer};
-use std::ptr::NonNull;
-
-pub unsafe fn deserialize<'de, TType, TLibfunc, D>(
-    deserializer: D,
-    arena: &Bump,
-    registry: &ProgramRegistry<TType, TLibfunc>,
-    info: &InfoAndTypeConcreteType,
-) -> Result<NonNull<()>, D::Error>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-    <TType as GenericType>::Concrete: TypeBuilder<TType, TLibfunc> + ValueBuilder<TType, TLibfunc>,
-    D: Deserializer<'de>,
-{
-    type ParamDeserializer<'a, TType, TLibfunc> =
-        <<TType as GenericType>::Concrete as ValueBuilder<TType, TLibfunc>>::Deserializer<'a>;
-
-    ParamDeserializer::<TType, TLibfunc>::new(arena, registry, registry.get_type(&info.ty).unwrap())
-        .deserialize(deserializer)
-}
-
-pub unsafe fn serialize<TType, TLibfunc, S>(
-    serializer: S,
-    registry: &ProgramRegistry<TType, TLibfunc>,
-    ptr: NonNull<()>,
-    info: &InfoAndTypeConcreteType,
-) -> Result<S::Ok, S::Error>
-where
-    TType: GenericType,
-    TLibfunc: GenericLibfunc,
-    <TType as GenericType>::Concrete: TypeBuilder<TType, TLibfunc> + ValueBuilder<TType, TLibfunc>,
-    S: Serializer,
-{
-    type ParamSerializer<'a, TType, TLibfunc> =
-        <<TType as GenericType>::Concrete as ValueBuilder<TType, TLibfunc>>::Serializer<'a>;
-
-    ParamSerializer::<TType, TLibfunc>::new(ptr, registry, registry.get_type(&info.ty).unwrap())
-        .serialize(serializer)
-}
